Extract common action invocation validator code to a new base class
