+++ src/Abp.AspNetCore/AspNetCore/Mvc/Validation/MvcActionInvocationValidator.cs
@@ -1,61 +1,36 @@
-﻿using System;
-using System.ComponentModel.DataAnnotations;
+﻿using System.ComponentModel.DataAnnotations;
 using Abp.AspNetCore.Mvc.Extensions;
 using Abp.Collections.Extensions;
 using Abp.Configuration.Startup;
 using Abp.Dependency;
-using Abp.Runtime.Validation.Interception;
+using Abp.Web.Validation;
 using Microsoft.AspNetCore.Mvc.Filters;
 
 namespace Abp.AspNetCore.Mvc.Validation
 {
-    public class MvcActionInvocationValidator : MethodInvocationValidator
+    public class MvcActionInvocationValidator : ActionInvocationValidatorBase
     {
         protected ActionExecutingContext ActionContext { get; private set; }
-
-        private bool _isValidatedBefore;
 
         public MvcActionInvocationValidator(IValidationConfiguration configuration, IIocResolver iocResolver)
             : base(configuration, iocResolver)
         {
-
         }
 
         public void Initialize(ActionExecutingContext actionContext)
         {
             ActionContext = actionContext;
 
-            base.Initialize(
-                actionContext.ActionDescriptor.GetMethodInfo(),
-                GetParameterValues(actionContext)
-            );
+            base.Initialize(actionContext.ActionDescriptor.GetMethodInfo());
         }
 
-        protected override bool ShouldValidateUsingValidator(object validatingObject, Type validatorType)
+        protected override object GetParameterValue(string parameterName)
         {
-            // Skip data annotations validation because MVC does this automatically
-            if (validatorType == typeof(DataAnnotationsValidator))
-            {
-                return false;
-            }
-
-            return base.ShouldValidateUsingValidator(validatingObject, validatorType);
+            return ActionContext.ActionArguments.GetOrDefault(parameterName);
         }
 
-        protected override void SetValidationErrors(object validatingObject)
+        protected override void SetDataAnnotationAttributeErrors()
         {
-            base.SetValidationErrors(validatingObject);
-
-            SetDataAnnotationAttributeErrors();
-        }
-
-        protected virtual void SetDataAnnotationAttributeErrors()
-        {
-            if (_isValidatedBefore || ActionContext.ModelState.IsValid)
-            {
-                return;
-            }
-
             foreach (var state in ActionContext.ModelState)
             {
                 foreach (var error in state.Value.Errors)
@@ -63,23 +38,6 @@ namespace Abp.AspNetCore.Mvc.Validation
                     ValidationErrors.Add(new ValidationResult(error.ErrorMessage, new[] { state.Key }));
                 }
             }
-
-            _isValidatedBefore = true;
-        }
-
-        protected virtual object[] GetParameterValues(ActionExecutingContext actionContext)
-        {
-            var methodInfo = actionContext.ActionDescriptor.GetMethodInfo();
-
-            var parameters = methodInfo.GetParameters();
-            var parameterValues = new object[parameters.Length];
-
-            for (var i = 0; i < parameters.Length; i++)
-            {
-                parameterValues[i] = actionContext.ActionArguments.GetOrDefault(parameters[i].Name);
-            }
-
-            return parameterValues;
         }
     }
 }
